backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
nibabel>=5.2.0
pydicom>=2.4.0
numpy>=1.26.0
imageio>=2.33.0
scikit-image>=0.22.0
matplotlib>=3.8.0
Pillow>=10.2.0
